Use attribute name instead of id where possible

+++ src/lib/server/repo/attribute.repo.ts
@@ -3,7 +3,7 @@ import { sql } from "$lib/server/sql";
 import { attributeValueRepo } from "./attribute-value.repo";
 
 export const attributeRepo = {
-	create: async (userId: number, name: string): Promise<Attribute | null> => {
+	create: async (userId: number, name: string): Promise<string | null> => {
 		const existingAttribute = await attributeRepo.getOneByName(userId, name);
 
 		if (existingAttribute) {
@@ -18,15 +18,17 @@ export const attributeRepo = {
 
 		console.info(`Created attribute ${attributes[0].id} for user ${userId}`);
 
-		return attributes[0];
+		return attributes[0].name;
 	},
 
-	getAll: async (userId: number): Promise<Attribute[]> => {
-		return await sql<Attribute[]>`
+	getAllNames: async (userId: number): Promise<string[]> => {
+		const attributes = await sql<Attribute[]>`
 			SELECT id, name, user_id
 			FROM attribute
 			WHERE user_id = ${userId}
 		`;
+
+		return attributes.map(nameOnly);
 	},
 
 	getOne: async (userId: number, attributeId: number): Promise<Attribute | null> => {
@@ -59,24 +61,36 @@ export const attributeRepo = {
 		return attributes;
 	},
 
-	update: async (userId: number, attributeId: number, name: string | null): Promise<void> => {
+	update: async (userId: number, attributeName: string, name: string | null): Promise<void> => {
 		await sql`
 			UPDATE attribute
 			SET name = COALESCE(${name}, name)
-			WHERE user_id = ${userId} AND id = ${attributeId}
+			WHERE user_id = ${userId} AND name = ${attributeName}
 		`;
 
-		console.info(`Updated attribute ${attributeId}`);
+		console.info(`Updated attribute ${attributeName}`);
 	},
 
-	delete: async (userId: number, attributeId: number): Promise<void> => {
-		attributeValueRepo.deleteByAttributeId(attributeId);
+	delete: async (userId: number, attributeName: string): Promise<boolean> => {
+		const attribute = await attributeRepo.getOneByName(userId, attributeName);
+
+		if (!attribute) {
+			return false;
+		}
+
+		await attributeValueRepo.deleteByAttributeId(attribute.id);
 
 		await sql`
 			DELETE FROM attribute
-			WHERE user_id = ${userId} AND id = ${attributeId}
+			WHERE user_id = ${userId} AND name = ${attributeName}
 		`;
 
-		console.info(`Deleted attribute ${attributeId}`);
+		console.info(`Deleted attribute ${attributeName}`);
+
+		return true;
 	},
 };
+
+function nameOnly(attribute: Attribute): string {
+	return attribute.name;
+}
